--- src/shared/components/common/Feedback/Notification.tsx
@@ -1,3 +1,0 @@
-import { Notification as MantineNotification } from '@mantine/core';
-
-export const Notification = MantineNotification;
