+++ src/shared/components/common/Feedback/Notifications.tsx
@@ -1,0 +1,3 @@
+import { notifications as MantineNotification } from '@mantine/notifications';
+
+export const notifications = MantineNotification;
