+++ src/shared/components/common/Feedback/index.ts
@@ -1,2 +1,2 @@
-export { Notification } from './Notification';
+export { notifications } from './Notifications';
 export { ErrorNotificationData } from './ErrorNotificationData';
